move ringbuffer to submodule
