+++ pyacq/core/stream/__init__.py
@@ -1,1 +1,2 @@
 from .stream import *
+from .sharedarray import SharedArray
